+++ app.js
@@ -12,6 +12,7 @@ const productRoutes = require('./routes/products');
 const userRoutes = require('./routes/users');
 const mailerRoutes = require('./routes/mailers');
 const orderRoutes = require('./routes/orders');
+const cartRoutes = require('./routes/carts');
 // const { celebrateCreateUser, celebrateLoginUser } = require('./validators/users');
 const NotFoundError = require('./errors/NotFoundError');
 const { requestLogger, errorLogger } = require('./middlewares/logger');
@@ -64,6 +65,8 @@ app.use(userRoutes);
 app.use(mailerRoutes);
 
 app.use(orderRoutes);
+
+app.use(cartRoutes);
 
 app.use((req, res, next) => next(new NotFoundError('Страница не найдена')));
 app.use(errorLogger);
