+++ routes/orders.js
@@ -3,7 +3,7 @@ const orderModel = require("../models/order");
 
 const router = express.Router();
 
-router.post("/create/order", async (req, res) => {
+router.post("/order/create", async (req, res) => {
   const orderData = req.body;
   try {
     const orderId = await orderModel.createOrder(orderData);
@@ -15,6 +15,24 @@ router.post("/create/order", async (req, res) => {
     }
   } catch (error) {
     console.error("Error creating order", error);
+    res.status(500).json({ error: "Internal Server Error" });
+  }
+});
+
+router.get("/order/:id", async (req, res) => {
+  const userId = req.params.id;
+
+  try {
+    const order = await orderModel.getOrdersByUserId(userId);
+
+    if (!order) {
+      res.status(404).json({ error: "Order not found" });
+      return;
+    }
+
+    res.json(order);
+  } catch (error) {
+    console.error("Error fetching order by ID:", error);
     res.status(500).json({ error: "Internal Server Error" });
   }
 });
